Rename 'pages' app to 'web' for clarity
