+++ templates/_base.html
@@ -19,8 +19,8 @@
         </a>
         <ul class="nav col-12 col-lg-auto me-lg-auto mb-2 justify-content-center mb-md-0">
           <li><a href="/" class="nav-link px-2 text-white">Home</a></li>
-          <li><a href="{% url 'pages:about' %}" class="nav-link px-2 text-white">About</a></li>
-          <li><a href="{% url 'pages:contact' %}" class="nav-link px-2 text-white">Contact</a></li>
+          <li><a href="{% url 'web:about' %}" class="nav-link px-2 text-white">About</a></li>
+          <li><a href="{% url 'web:contact' %}" class="nav-link px-2 text-white">Contact</a></li>
         </ul>
         {% if user.is_authenticated %}
         <a href="{% url 'leagues:league_list' %}" class="btn btn-outline-light me-lg-3">
