+++ templates/leagues/_league_base.html
@@ -101,7 +101,7 @@
                   </button>
                   <div class="collapse show" id="account-collapse">
                     <ul class="btn-toggle-nav list-unstyled fw-normal pb-1 small">
-                      <li><a href="{% url 'pages:home' %}" class="link-dark rounded">Exit Game</a></li>
+                      <li><a href="{% url 'web:home' %}" class="link-dark rounded">Exit Game</a></li>
                       <li><a href="{% url 'account_change_password' %}" class="link-dark rounded">Change Password</a></li>
                       <li><a href="{% url 'account_logout' %}" class="link-dark rounded">Sign out</a></li>                      
                     </ul>
