+++ templates/leagues/league/league_list.html
@@ -4,7 +4,7 @@
 
 {% block content %}
   <h2>My Leagues &middot; {{ user }}</h2>
-  <p><a href="{% url 'pages:home' %}">&larr; Home</a></p>
+  <p><a href="{% url 'web:home' %}">&larr; Home</a></p>
   <hr>
   <a href="{% url 'leagues:league_create' %}">
     <button class="btn btn-dark mt-2 mb-3" type="submit">New League</button>
